# docs/swagger.yaml
basePath: /
definitions:
...
  utils.DatabaseSizePayload:
    properties:
      size_string:
        type: integer
    type: object
...
host: localhost:8101
info:
  contact: {}
  description: REST API for monitoring service uptime, response times, and managing
    configuration.
  title: goUp API
  version: "1.0"
paths:
...
  /api/db/clear:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "500":
          description: internal server error
          schema:
            type: string
      summary: Clear all data from the database
      tags:
      - database
  /api/db/persist:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: persistence enabled
          schema:
            type: boolean
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get whether database persistence is enabled
      tags:
      - database
    post:
      produces:
      - application/json
      responses:
        "200":
          description: empty body
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Toggle database persistence
      tags:
      - database
  /api/db/size:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DatabaseSizePayload'
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get database file size in bytes for regular files; system-dependent
        for others https://pkg.go.dev/io/fs#FileInfo.Size
      tags:
      - database
...
schemes:
- http
swagger: "2.0"